requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
google-cloud-aiplatform==1.89.0
python-dotenv==1.0.0 